+++ src/client.ts
@@ -21,6 +21,8 @@ import {createCommandStop} from "#commands/CommandStop/CommandStop";
 import {createCommandRestart} from "#commands/CommandRestart/CommandRestart";
 import {createCommandUpdate} from "#commands/CommandUpdate/CommandUpdate";
 import {createCommandIdentity} from "#commands/CommandIdentity/CommandIdentity";
+import {createCommandShare} from "#commands/CommandShare/CommandShare";
+import {createCommandConnect} from "#commands/CommandConnect/CommandConnect";
 import {ApiClient} from "#api-client/ApiClient";
 import {readPackageJson} from "#package-json/packageJson";
 
@@ -67,6 +69,8 @@ program.addCommand(createCommandStop(ctx, program))
 program.addCommand(createCommandRestart(ctx, program))
 program.addCommand(createCommandUpdate(ctx, program))
 program.addCommand(createCommandIdentity(ctx, program))
+program.addCommand(createCommandShare(ctx, program))
+program.addCommand(createCommandConnect(ctx, program))
 
 program.action(({options}: ParseResult) => {
   if (options.version) {